requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
Pillow
seaborn
tqdm
scikit-image
timm